refactor(snowman): relocate impl to src/snowman/index.ts behind a thin facade (R3.7)
First R3 step. `git mv src/snowman.ts src/snowman/index.ts` (the whole
implementation — `Snowman` + the contract types — moves wholesale), and add a thin
`src/snowman.ts` facade that does `export * from './snowman/index.js'`. Every
`./snowman.js` importer (main.ts side-effect, snowglider.ts, game/*.ts, physics.ts
types, ui/hud.ts types, browser-tests.js, contract-surface-tests.js) keeps resolving
a sibling file, so the import seam is invisible. Nothing is decomposed yet — steps
8–12 carve model/pose/physics/collision/test-hooks out of index.ts.

Resolution gotcha handled (per the plan): the physics-invariant harness runs under
bare Node with no `.js`->`.ts` resolve hook, so it cannot resolve the facade's
`./snowman/index.js` re-export. Repointed it to import the relocated implementation
directly (`src/snowman/index.ts`); `index.ts` stays free of relative imports so it
loads under bare Node. `test:verify` stays hook-free (no package.json change).

Verification: typecheck / lint (0 errors) / build green; **test:verify physics-
invariant harness OK (coasting byte-identical to snowman_baseline.js)** + dom smoke
18/0; contract-surface guard 44/44 (the `./snowman.js` runtime seam resolves Snowman
+ all 4 methods); puppeteer 87/87 across 7/7 suites.
