+++ docs/ARCHITECTURE.md
@@ -234,7 +234,15 @@ camera.position -= shake                          // revert so smoothing stays c
 > `createMainLoop()` / `createLifecycle()` wiring, the eager `Snowman.addTestHooks`
 > calls, `window.initializeGameWithAudio`, `publishGameGlobals()`, and the test-mode
 > auto-start. (The dead local `addTestHooks` shim was deleted — the real browser hooks
-> come from `Snowman.addTestHooks`.) `snowman.ts` is the next target (Stage R3).
+> come from `Snowman.addTestHooks`.)
+>
+> Stage R3 (in progress) splits `snowman.ts`: the implementation moved to
+> **`src/snowman/index.ts`** behind a thin **`src/snowman.ts`** facade
+> (`export * from './snowman/index.js'`) so every `./snowman.js` importer keeps
+> resolving a sibling file. `index.ts` stays free of *relative* imports so the
+> bare-Node physics-invariant harness can load it directly (it imports
+> `src/snowman/index.ts`, not the facade). Steps 8–12 carve `model` / `pose` /
+> `physics` / `collision` / `test-hooks` out of `index.ts`.
 
 ---
