+++ src/snowman.ts
@@ -1,1079 +1,23 @@
-// snowman.ts - Snowman model and functions for SnowGlider game
+// snowman.ts — ROOT FACADE for the Snowman module (Stage R3, issue #34).
 //
-// Phase 2.8 (issue #84): final terrain-cluster module converted off the classic
-// global model. `THREE` now comes from the npm package via a real ES-module
-// import instead of the CDN global, and `Snowman` is `export`ed. snowman.js
-// receives the terrain samplers (getTerrainHeight/getTerrainGradient/
-// getDownhillDirection) as function arguments (not globals); it is loaded into the
-// page through the bundle entry (src/main.js) and imported by snowglider.js.
+// The implementation moved wholesale to `src/snowman/index.ts`; this thin file
+// re-exports it so the existing importers keep resolving a *sibling* `./snowman.js`
+// specifier (there is no directory-index resolution here):
+//   - src/main.ts            (side-effect import, keeps it in the bundle graph)
+//   - src/snowglider.ts      (import { Snowman })
+//   - src/game/*.ts          (import { Snowman })
+//   - src/physics.ts         (imports the contract *types*)
+//   - src/ui/hud.ts          (imports PlayerPos / UpdateResult types)
+//   - tests/browser-tests.js (import { Snowman } from '../src/snowman.js')
+//   - tests/contract-surface-tests.js (runtime import, via the .js->.ts hook)
 //
-// Phase 3.7 (issue #84): renamed `.js` -> `.ts`. The `@ts-check` pragma is gone
-// (implied for a real `.ts` file) and the movement/physics contract is now
-// expressed as real type declarations: the player position/velocity inputs, the
-// injected terrain samplers, the controls object, the tree-collision shape, the
-// camera-manager seam, and the per-frame `UpdateResult`. The physics math is
-// byte-identical — every edit is type-only/erasable, so esbuild (Vite) and Node's
-// native type-stripping run it exactly as before; the physics-invariant harness
-// confirms coasting stays bit-identical to the frozen baseline.
-import * as THREE from 'three';
-
-// These contract types are exported so the typed player-state layer in
-// physics.ts (PR 3.21) shares snowman's exact call contract instead of
-// re-declaring it. Exporting interfaces/types is purely additive and erasable.
-
-/** Mutable player position the physics integrates each frame. */
-export interface PlayerPos {
-  x: number;
-  y: number;
-  z: number;
-}
-
-/** Mutable horizontal velocity (vertical motion is tracked by verticalVelocity). */
-export interface PlanarVelocity {
-  x: number;
-  z: number;
-}
-
-/** A 2D terrain vector: a gradient or a unit downhill direction. */
-interface TerrainVec2 {
-  x: number;
-  z: number;
-}
-
-/** Terrain height sampler injected by the orchestrator. */
-export type TerrainHeightFn = (x: number, z: number) => number;
-/** Terrain gradient / downhill-direction sampler injected by the orchestrator. */
-export type TerrainVecFn = (x: number, z: number) => TerrainVec2;
-
-/** The control flags updateSnowman reads each frame. */
-export interface SnowmanControls {
-  left: boolean;
-  right: boolean;
-  up: boolean;
-  down: boolean;
-  jump: boolean;
-}
-
-/** Minimal tree-position shape the collision check reads. */
-export interface TreePos {
-  x: number;
-  y: number;
-  z: number;
-}
-
-/** Minimal rock-position shape the collision check reads. */
-export interface RockPos {
-  x: number;
-  y: number;
-  z: number;
-  size: number;
-}
-
-/** The camera-manager seam resetSnowman drives (satisfied by the Camera class). */
-export interface CameraManagerLike {
-  initialize(position: THREE.Vector3, rotation: THREE.Euler): void;
-}
-
-/** Game-over callback handed in by the orchestrator. */
-export type ShowGameOverFn = (reason: string) => void;
-
-/** Ski technique surfaced for the HUD + ski pose. */
-type SkiTechnique = 'air' | 'glide' | 'snowplow' | 'skid' | 'carve' | 'parallel' | 'tuck' | 'hop';
-
-/** Per-frame physics output returned by updateSnowman. */
-export interface UpdateResult {
-  isInAir: boolean;
-  verticalVelocity: number;
-  lastTerrainHeight: number;
-  airTime: number;
-  jumpCooldown: number;
-  turnPhase: number;
-  currentTurnDirection: number;
-  turnChangeCooldown: number;
-  currentSpeed: number;
-  technique: SkiTechnique;
-  justLanded: boolean;
-  landingForce: number;
-}
-
-// Create Snowman (Three Spheres)
-function createSnowman(scene: THREE.Scene): THREE.Group {
-  const group = new THREE.Group();
-  const snowMaterial = new THREE.MeshStandardMaterial({ color: 0xffffff, roughness: 0.9 });
-  const blackMaterial = new THREE.MeshStandardMaterial({ color: 0x000000 });
-  const carrotMaterial = new THREE.MeshStandardMaterial({ color: 0xFF6600 });
-  const stickMaterial = new THREE.MeshStandardMaterial({ color: 0x8B4513 }); // Brown for sticks
-  const hatMaterial = new THREE.MeshStandardMaterial({ color: 0x111111 }); // Dark grey/black for hat
-  
-  // Bottom sphere
-  const bottom = new THREE.Mesh(new THREE.SphereGeometry(2, 24, 24), snowMaterial);
-  bottom.position.y = 2;
-  bottom.castShadow = true;
-  group.add(bottom);
-  
-  // Middle sphere
-  const middle = new THREE.Mesh(new THREE.SphereGeometry(1.5, 24, 24), snowMaterial);
-  middle.position.y = 4.5;
-  middle.castShadow = true;
-  group.add(middle);
-  
-  // Head sphere
-  const head = new THREE.Mesh(new THREE.SphereGeometry(1, 24, 24), snowMaterial);
-  head.position.y = 7.0; // Lowered head to sit on middle sphere (4.5 + 1.5 + 1.0)
-  head.castShadow = true;
-  group.add(head);
-  
-  // Eyes
-  const leftEye = new THREE.Mesh(new THREE.SphereGeometry(0.15, 12, 12), blackMaterial);
-  leftEye.position.set(0.4, 7.2, 0.8); // Adjusted Y based on new head position
-  group.add(leftEye);
-  
-  const rightEye = new THREE.Mesh(new THREE.SphereGeometry(0.15, 12, 12), blackMaterial);
-  rightEye.position.set(-0.4, 7.2, 0.8); // Adjusted Y based on new head position
-  group.add(rightEye);
-  
-  // Carrot nose
-  const nose = new THREE.Mesh(new THREE.ConeGeometry(0.2, 1, 12), carrotMaterial);
-  nose.position.set(0, 7.0, 1); // Adjusted Y based on new head position
-  nose.rotation.x = Math.PI / 2;
-  group.add(nose);
-  
-  // Buttons (on middle and bottom spheres)
-  const buttonGeometry = new THREE.SphereGeometry(0.15, 12, 12);
-  const button1 = new THREE.Mesh(buttonGeometry, blackMaterial);
-  button1.position.set(0, 5.5, 1.4); // On middle sphere front
-  group.add(button1);
-  
-  const button2 = new THREE.Mesh(buttonGeometry, blackMaterial);
-  button2.position.set(0, 4.5, 1.45); // On middle sphere front
-  group.add(button2);
-  
-  const button3 = new THREE.Mesh(buttonGeometry, blackMaterial);
-  button3.position.set(0, 3.0, 1.9); // On bottom sphere front
-  group.add(button3);
-  
-  // --- Stick Arms ---
-  // Create a function to build a branched arm
-  function createBranchArm(isLeft: boolean): THREE.Group {
-    const armGroup = new THREE.Group();
-    const mainStickLength = 2.5;
-    const mainStickRadius = 0.08; // Slightly thinner radius
-    const segments = 8; // Segments for the tube
-
-    // Create a slightly irregular path for the main stick
-    const pathPoints = [
-      new THREE.Vector3(0, 0, 0),
-      new THREE.Vector3(0, mainStickLength * 0.4, isLeft ? 0.1 : -0.1), // Slight bend
-      new THREE.Vector3(isLeft ? -0.1 : 0.1, mainStickLength * 0.8, 0), // Another slight bend
-      new THREE.Vector3(0, mainStickLength, 0) // End point
-    ];
-    const curve = new THREE.CatmullRomCurve3(pathPoints);
-
-    // Main stick using TubeGeometry
-    const mainStickGeom = new THREE.TubeGeometry(curve, segments * 2, mainStickRadius, 5, false); // Fewer radial segments
-    const mainStick = new THREE.Mesh(mainStickGeom, stickMaterial);
-    // No need to translate geometry origin with TubeGeometry path starting at 0,0,0
-    mainStick.rotation.z = isLeft ? -Math.PI / 10 : Math.PI / 10; // Adjusted angle slightly
-    mainStick.castShadow = true;
-    armGroup.add(mainStick);
-
-    // Small branch 1 (thinner)
-    const branch1Length = 0.8;
-    const branch1Radius = 0.05;
-    const branch1 = new THREE.Mesh(
-      new THREE.CylinderGeometry(branch1Radius, branch1Radius * 0.7, branch1Length, 5), // Fewer segments
-      stickMaterial
-    );
-    branch1.geometry.translate(0, branch1Length / 2, 0); // Move origin to base
-    // Attach near the first bend point using curve.getPointAt
-    const attachPoint1 = curve.getPointAt(0.5); // Get point halfway along the curve
-    branch1.position.copy(attachPoint1);
-    branch1.rotation.z = isLeft ? -Math.PI / 5 : Math.PI / 5; // Angle outward more
-    branch1.rotation.x = Math.PI / 12; // Angle slightly forward
-    branch1.castShadow = true;
-    mainStick.add(branch1); // Add as child of main stick
-
-    // Small branch 2 (optional, smaller and thinner)
-    const branch2Length = 0.5;
-    const branch2Radius = 0.04;
-    const branch2 = new THREE.Mesh(
-      new THREE.CylinderGeometry(branch2Radius, branch2Radius * 0.7, branch2Length, 5), // Fewer segments
-      stickMaterial
-    );
-    branch2.geometry.translate(0, branch2Length / 2, 0); // Move origin to base
-    // Attach near the second bend point
-    const attachPoint2 = curve.getPointAt(0.8); // Get point further along the curve
-    branch2.position.copy(attachPoint2);
-    branch2.rotation.z = isLeft ? Math.PI / 4 : -Math.PI / 4; // Angle more sharply
-    branch2.rotation.x = -Math.PI / 10; // Angle slightly backward
-    branch2.castShadow = true;
-    mainStick.add(branch2); // Add as child of main stick
-
-    // Position the entire arm group
-    // Attach to middle sphere, slightly adjusted position
-    armGroup.position.set(isLeft ? 1.35 : -1.35, 4.9, 0); // Adjusted position slightly
-    // Rotate slightly forward and outward
-    armGroup.rotation.x = Math.PI / 16; // Less forward tilt
-    armGroup.rotation.y = isLeft ? -Math.PI / 8 : Math.PI / 8; // Rotate arms outward more
-
-    return armGroup;
-  }
-
-  // Left Arm
-  const leftArmGroup = createBranchArm(true);
-  group.add(leftArmGroup);
-
-  // Right Arm
-  const rightArmGroup = createBranchArm(false);
-  group.add(rightArmGroup);
-  // --- End Stick Arms ---
-  
-  // Hat
-  const hatBase = new THREE.Mesh(new THREE.CylinderGeometry(0.7, 0.7, 0.2, 24), hatMaterial);
-  hatBase.position.y = 7.9; // Adjusted Y based on new head position: 7.0 (head_y) + 1.0 (head_r) - 0.1
-  hatBase.castShadow = true;
-  group.add(hatBase);
-  
-  const hatTop = new THREE.Mesh(new THREE.CylinderGeometry(0.45, 0.45, 0.9, 24), hatMaterial);
-  hatTop.position.y = 8.45; // Adjusted Y: 7.9 (base_y) + 0.1 (base_half_h) + 0.45 (new_top_half_h)
-  hatTop.castShadow = true;
-  group.add(hatTop);
-  
-  // Add skis
-  const skiMaterial = new THREE.MeshStandardMaterial({ color: 0xFF0000 }); // Bright red
-  
-  // Left ski
-  const leftSki = new THREE.Mesh(
-    new THREE.BoxGeometry(0.8, 0.2, 6), 
-    skiMaterial
-  );
-  leftSki.position.set(-1, 0.1, 1);
-  leftSki.castShadow = true;
-  // Add ski tip (angled front)
-  const leftSkiTip = new THREE.Mesh(
-    new THREE.BoxGeometry(0.8, 0.4, 1),
-    skiMaterial
-  );
-  leftSkiTip.position.set(0, 0.2, 3);
-  leftSkiTip.rotation.x = Math.PI / 8; // Angle up slightly
-  leftSki.add(leftSkiTip);
-  group.add(leftSki);
-  
-  // Right ski
-  const rightSki = new THREE.Mesh(
-    new THREE.BoxGeometry(0.8, 0.2, 6),
-    skiMaterial
-  );
-  rightSki.position.set(1, 0.1, 1);
-  rightSki.castShadow = true;
-  // Add ski tip (angled front)
-  const rightSkiTip = new THREE.Mesh(
-    new THREE.BoxGeometry(0.8, 0.4, 1),
-    skiMaterial
-  );
-  rightSkiTip.position.set(0, 0.2, 3);
-  rightSkiTip.rotation.x = Math.PI / 8; // Angle up slightly
-  rightSki.add(rightSkiTip);
-  group.add(rightSki);
-  
-  // Keep references + neutral pose so ski technique (e.g. snowplow wedge) can be shown.
-  group.userData = group.userData || {};
-  group.userData.leftSki = leftSki;
-  group.userData.rightSki = rightSki;
-  group.userData.leftSkiBaseX = leftSki.position.x;
-  group.userData.rightSkiBaseX = rightSki.position.x;
-  
-  scene.add(group);
-  return group;
-}
-
-// Reset the snowman to initial position
-function resetSnowman(snowman: THREE.Object3D, pos: PlayerPos, velocity: PlanarVelocity, getTerrainHeight: TerrainHeightFn, cameraManager: CameraManagerLike): number {
-  // Start higher up the mountain (z=-20 instead of -40 for a longer run)
-  // With extended terrain, we can start even higher up at z=-15
-  pos.x = 0;
-  pos.z = -15;
-  pos.y = getTerrainHeight(0, -15);
-  
-  // Smoother initial velocity - reduce the starting velocity to avoid the initial jolt
-  velocity.x = 0;
-  velocity.z = -3.0;
-  
-  // Reset user data for smooth motion if it exists
-  if (snowman.userData) {
-    snowman.userData.targetRotationY = Math.PI; // Default facing downhill
-    snowman.userData.currentRotX = 0;
-    snowman.userData.currentRotZ = 0;
-    // Clear edge-engagement state so a new run starts with no locked carve.
-    snowman.userData.carveCharge = 0;
-    snowman.userData.lastSteerDir = 0;
-  }
-  
-  // Force all rotations to be explicit - avoid any chance of NaN or unexpected values
-  snowman.position.set(pos.x, pos.y, pos.z);
-  snowman.rotation.set(0, Math.PI, 0);
-  
-  // Explicitly reset the rotation tracking for snowman tilt
-  snowman.rotation.x = 0;
-  snowman.rotation.z = 0;
-  
-  // Reset camera using the camera manager
-  cameraManager.initialize(snowman.position, snowman.rotation);
-  
-  return getTerrainHeight(0, -15); // Return the terrain height for lastTerrainHeight
-}
-
-// Update snowman physics and movement
-function updateSnowman(snowman: THREE.Object3D, delta: number, pos: PlayerPos, velocity: PlanarVelocity, isInAir: boolean, verticalVelocity: number,
-                      lastTerrainHeight: number, airTime: number, jumpCooldown: number, controls: SnowmanControls,
-                      turnPhase: number, currentTurnDirection: number, turnChangeCooldown: number, turnAmplitude: number,
-                      getTerrainHeight: TerrainHeightFn, getTerrainGradient: TerrainVecFn, getDownhillDirection: TerrainVecFn,
-                      treePositions: TreePos[], gameActive: boolean, showGameOver: ShowGameOverFn,
-                      rockPositions: RockPos[] = []): UpdateResult {
-  
-  // Update jump cooldown
-  if (jumpCooldown > 0) {
-    jumpCooldown -= delta;
-  }
-  
-  // Outer-scope outputs surfaced in the return value (HUD + camera juice).
-  let technique: SkiTechnique = isInAir ? 'air' : 'glide';
-  let justLanded = false;
-  let landingForce = 0;
-  
-  // Get current terrain height at position
-  const terrainHeightAtPosition = getTerrainHeight(pos.x, pos.z);
-  
-  // Check for landing
-  if (isInAir && pos.y <= terrainHeightAtPosition) {
-    isInAir = false;
-    pos.y = terrainHeightAtPosition;
-    justLanded = true;
-    
-    // Landing impact based on air time and height
-    const landingImpact = Math.min(0.5, airTime * 0.15);
-    landingForce = airTime; // seconds aloft; used for camera shake on touchdown
-    const currentSpeed = Math.sqrt(velocity.x*velocity.x + velocity.z*velocity.z);
-    
-    // Reduce speed on landing
-    velocity.x *= (1 - landingImpact);
-    velocity.z *= (1 - landingImpact);
-    
-    // Reset jump-related variables
-    verticalVelocity = 0;
-    airTime = 0;
-    jumpCooldown = 0.3; // Short cooldown after landing
-  }
-  
-  // Calculate the downhill direction
-  const dir = getDownhillDirection(pos.x, pos.z);
-  
-  // Get gradient for physics calculations
-  const gradient = getTerrainGradient(pos.x, pos.z);
-  const steepness = Math.sqrt(gradient.x*gradient.x + gradient.z*gradient.z);
-  
-  // Detect natural jumps from terrain (like going over moguls)
-  const heightDifference = terrainHeightAtPosition - lastTerrainHeight;
-  const currentSpeed = Math.sqrt(velocity.x*velocity.x + velocity.z*velocity.z);
-  const movingFast = currentSpeed > 12;
-  
-  // Auto-jump when going downhill after a steep uphill section
-  if (!isInAir && heightDifference < -0.8 && movingFast && jumpCooldown <= 0) {
-    verticalVelocity = 6 + (currentSpeed * 0.3);
-    isInAir = true;
-  }
-  
-  // Manual jump / hop turn with spacebar or touch (grounded, off cooldown).
-  // Plain Jump = a straight pop into the air. Jump WHILE steering Left/Right = a
-  // hop turn (issue #48): a quick edge-set pivot that snaps the heading toward the
-  // steer direction and scrubs speed — the steep-terrain "hop the skis around and
-  // set them down pointing the new way" move. It trades speed (HOP_SPEED_KEEP < 1)
-  // for a sharper direction change than carving can give, and lands you on a fresh
-  // edge committed to the new direction (carveCharge reset, lastSteerDir set). It
-  // is fully gated behind explicit jump+steer input, so the no-input invariant and
-  // every plain-steering harness check are untouched.
-  if (controls.jump && !isInAir && jumpCooldown <= 0) {
-    const hopSteer = (controls.left ? -1 : 0) + (controls.right ? 1 : 0);
-    if (hopSteer !== 0) {
-      const HOP_PIVOT_ANGLE = 0.4; // rad (~23°) the velocity heading snaps per hop
-      const HOP_SPEED_KEEP = 0.82; // a hop turn scrubs ~18% of horizontal speed
-      const HOP_POP = 5.0;         // small vertical pop, well below a full jump
-      const HOP_COOLDOWN = 0.45;   // s; prevents hop spam
-      // Rotate the horizontal velocity toward the steer direction (right => +x).
-      const theta = hopSteer * HOP_PIVOT_ANGLE;
-      const c = Math.cos(theta), s = Math.sin(theta);
-      const nvx = velocity.x * c - velocity.z * s;
-      const nvz = velocity.x * s + velocity.z * c;
-      velocity.x = nvx * HOP_SPEED_KEEP;
-      velocity.z = nvz * HOP_SPEED_KEEP;
-      verticalVelocity = HOP_POP;
-      isInAir = true;
-      jumpCooldown = HOP_COOLDOWN;
-      // Land on a fresh edge committed to the new direction.
-      if (snowman.userData) {
-        snowman.userData.carveCharge = 0;
-        snowman.userData.lastSteerDir = hopSteer;
-      }
-      technique = 'hop';
-    } else {
-      verticalVelocity = 10 + (currentSpeed * 0.5);
-      isInAir = true;
-      jumpCooldown = 0.5; // Prevent jump spam
-    }
-  }
-  
-  // Update vertical position and velocity when in air
-  if (isInAir) {
-    // Track time in air
-    airTime += delta;
-    
-    // Apply gravity to vertical velocity
-    verticalVelocity -= 16 * delta;
-    
-    // Update vertical position
-    pos.y += verticalVelocity * delta;
-    
-    // Air control
-    if (controls.left) {
-      velocity.x -= 5.0 * delta;
-    }
-    if (controls.right) {
-      velocity.x += 5.0 * delta;
-    }
-    
-    // Less friction in air
-    velocity.x *= (1 - 0.01);
-    velocity.z *= (1 - 0.01);
-  } else {
-    // Update velocity based on gravity, gradient, and an improved friction model
-    const gravity = 9.8;
-    
-    // Dynamic friction based on speed - less friction at lower speeds for smoother acceleration
-    // This prevents the jittery start by reducing initial resistance
-    const speedFactor = Math.min(1, currentSpeed / 8);
-    const baseFriction = 0.015; // Lower base friction for smoother starts
-    const friction = baseFriction + (0.025 * speedFactor); // Maximum 0.04 at high speeds
-    
-    // Apply forces to velocity (gravity pulls along slope direction) with smoother acceleration
-    velocity.x += dir.x * steepness * gravity * delta;
-    velocity.z += dir.z * steepness * gravity * delta;
-    
-    // --- Ski technique model -------------------------------------------------
-    // Layered on top of the original arcade handling. Crucially, when the player
-    // gives NO steering or brake input the behaviour below is identical to the
-    // original (turnForce/accel unchanged, skidScrub == 0), so coasting physics
-    // and the existing test expectations are preserved. Skill only emerges once
-    // the player actually works the edges:
-    //   - Snowplow (brake / Down): sheds real speed but grants tight, planted
-    //     turns ("pizza" stop) — slow and controllable.
-    //   - Carving (Left/Right): smooth, anticipatory turns hold speed; sharp
-    //     direction changes at speed wash the edges out and scrub speed (skid).
-    //   - Tuck / straight-line (Up, no steer): least friction, most speed, least
-    //     room to react — the risk/reward line.
-    const steering = (controls.left ? -1 : 0) + (controls.right ? 1 : 0);
-    const snowplow = !!controls.down && !isInAir;
-
-    // Terrain-dependent grip: a touch more bite on moderate pitches, looser when flat.
-    const terrainGrip = 0.6 + Math.min(0.4, steepness * 0.5);
-
-    // Steering authority: snowplow tightens the turn, speed loosens it slightly.
-    let turnForce = 16.0;
-    if (snowplow) turnForce = 24.0;            // planted wedge = sharper steering
-    else if (currentSpeed > 18) turnForce = 14.0; // hard to wrench the skis at speed
-
-    if (controls.left) {
-      velocity.x -= turnForce * delta;
-    }
-    if (controls.right) {
-      velocity.x += turnForce * delta;
-    }
-
-    // Forward input / straight-line tuck.
-    const accelerationForce = 10.0;
-    if (controls.up) {
-      velocity.z -= accelerationForce * delta;
-    }
-
-    // Snowplow braking: decelerate along the actual direction of travel so it
-    // bleeds genuine speed (not just downhill velocity), with a little extra dig.
-    // Clamp the impulse to the current speed so braking can bring the snowman to a
-    // stop but never reverse the velocity vector — otherwise at low speed the
-    // subtraction overshoots zero and the control bias below drives it back uphill,
-    // letting players climb/stall the timed course by braking.
-    if (snowplow && currentSpeed > 0.001) {
-      const brakeDecel = 14.0;
-      const brakeImpulse = Math.min(brakeDecel * delta, currentSpeed);
-      velocity.x -= (velocity.x / currentSpeed) * brakeImpulse;
-      velocity.z -= (velocity.z / currentSpeed) * brakeImpulse;
-      // Only nudge the slight uphill control bias while still moving; never after the
-      // brake has stopped the snowman (that would push it uphill from a standstill).
-      if (brakeImpulse < currentSpeed) {
-        velocity.z += accelerationForce * delta * 0.3;
-      }
-    }
-
-    // --- Edge engagement: carve vs skid (issues #48 / #54) -------------------
-    // Turning is only a real *skill* if a clean, committed turn costs less speed
-    // than a panicked one. `carveCharge` (0..1) tracks how "locked in" the
-    // current edge is: it builds while the player commits to ONE steering
-    // direction and collapses to 0 the instant they reverse or first set an edge
-    // from straight. A locked carve sheds little speed; a fresh or reversed edge
-    // washes out sideways and scrubs hard. So anticipating and holding a smooth
-    // line keeps speed, while flip-flopping the skis bleeds it — the
-    // speed-management trade-off the roadmap calls for.
-    //
-    // The state lives on snowman.userData (like `technique` and the pose state)
-    // so it persists across frames without widening the updateSnowman contract,
-    // and resetSnowman clears it between runs. Crucially it is *read/written but
-    // never used to alter velocity when the player gives no steering input*, so
-    // the no-input coasting path stays byte-identical to the frozen baseline.
-    const CARVE_BUILD_RATE = 1.5;    // ~0.66s of a held turn to lock the carve in
-    const CARVE_RELEASE_RATE = 3.0;  // edge releases ~2x faster than it engages
-    const CARVE_SCRUB_RELIEF = 0.8;  // a locked carve sheds up to 80% of edge wash-out
-    const TURN_TAX = 0.01;           // small always-on turn cost so a turn is never free
-    const PARALLEL_LOCK = 0.85;      // carveCharge past this = a mastered parallel turn
-
-    const ud = snowman.userData || (snowman.userData = {});
-    let carveCharge = ud.carveCharge || 0;
-    let lastSteerDir = ud.lastSteerDir || 0;
-    if (steering !== 0) {
-      // Same direction as last frame => the edge keeps engaging; a reversal or a
-      // fresh edge out of a straight line breaks it and starts the carve over.
-      carveCharge = steering === lastSteerDir
-        ? Math.min(1, carveCharge + delta * CARVE_BUILD_RATE)
-        : 0;
-      lastSteerDir = steering;
-    } else {
-      carveCharge = Math.max(0, carveCharge - delta * CARVE_RELEASE_RATE);
-      lastSteerDir = 0;
-    }
-    ud.carveCharge = carveCharge;
-    ud.lastSteerDir = lastSteerDir;
-
-    // Edge skid / carve drag: only meaningful while steering. Snowplow adds grip,
-    // so braking through a turn stays controlled instead of washing out.
-    let skidScrub = 0;
-    if (steering !== 0 && currentSpeed > 4) {
-      const speedFactor2 = Math.min(1, currentSpeed / 22);
-      const grip = snowplow ? 1.0 : terrainGrip;
-      // Edge wash-out: sharper turn relative to grip => more scrub (~0..0.06). A
-      // locked carve (carveCharge→1) sheds most of it; a skid (carveCharge→0)
-      // keeps all of it.
-      const edgeScrub = 0.06 * speedFactor2 * (1 - grip * 0.85) * (1 - CARVE_SCRUB_RELIEF * carveCharge);
-      // Plus an always-on turn tax (scaled by speed) so committing to a turn still
-      // costs a little versus straight-lining — turning is never free, but a clean
-      // carve is cheap. A fully-locked PARALLEL turn (carveCharge past the 0.85
-      // parallel threshold — skis together and rolled onto edge) is the most
-      // efficient turn there is, so the tax fades out across the parallel lock:
-      // mastering the carve into a parallel turn makes it nearly free. Below 0.85
-      // `parallelLock` is 0, so carve/skid feel and the gating carve-vs-skid check
-      // are byte-identical.
-      const parallelLock = Math.min(1, Math.max(0, (carveCharge - PARALLEL_LOCK) / (1 - PARALLEL_LOCK)));
-      skidScrub = edgeScrub + TURN_TAX * speedFactor2 * (1 - parallelLock);
-    }
-
-    // Expose technique for HUD + ski pose. A turn reads as a "skid" until the edge
-    // locks in (carveCharge), then a "carve", and finally a "parallel" turn once
-    // the edge is fully committed past PARALLEL_LOCK — the mastery tier above the
-    // beginner snowplow wedge.
-    technique = 'glide';
-    if (isInAir) technique = 'air';
-    else if (snowplow) technique = 'snowplow';
-    else if (steering !== 0) technique = carveCharge > PARALLEL_LOCK ? 'parallel' : (carveCharge > 0.55 ? 'carve' : 'skid');
-    else if (controls.up) technique = 'tuck';
-    
-    // Only use automatic turning if no user input
-    if (!controls.left && !controls.right) {
-      // Update turn phase and apply automatic turning
-      turnPhase += delta * 0.5; // Slower phase advancement for gentler turning
-      turnChangeCooldown -= delta;
-      
-      // More gradual turn direction changes
-      if (turnChangeCooldown <= 0) {
-        // Instead of completely random direction, bias toward centered movement
-        // Higher probability of returning to center line when far from center
-        const centeringBias = Math.min(1, Math.abs(pos.x) / 20) * 0.7;
-        const randomFactor = Math.random();
-        
-        if (pos.x > 5 && randomFactor < (0.6 + centeringBias)) {
-          // If we're right of center, bias toward turning left
-          currentTurnDirection = -1;
-        } else if (pos.x < -5 && randomFactor < (0.6 + centeringBias)) {
-          // If we're left of center, bias toward turning right
-          currentTurnDirection = 1;
-        } else {
-          // Otherwise random direction with smooth transition
-          currentTurnDirection = Math.random() > 0.5 ? 1 : -1;
-        }
-        
-        // Longer cooldown for smoother movement
-        turnChangeCooldown = 3 + Math.random() * 2;
-      }
-      
-      // Scale turn intensity with speed, but with much gentler effect at low speeds
-      const turnIntensity = Math.min(currentSpeed, 10) / 10;
-      
-      // Use eased sine wave for smoother turning - using 0.3 for smoother transition
-      velocity.x += Math.sin(turnPhase * 0.3) * (turnAmplitude * 0.7) * delta * turnIntensity * currentTurnDirection;
-    }
-    
-    // Apply friction to slow down. Base friction is unchanged when not steering
-    // (skidScrub == 0), so straight-line/coasting behaviour is identical to before;
-    // hard turns at speed add edge-skid drag on top.
-    const totalFriction = friction + skidScrub;
-    velocity.x *= (1 - totalFriction);
-    velocity.z *= (1 - totalFriction);
-    
-    // Show the current technique on the snowman: a snowplow forms a beginner ski
-    // wedge (tips inward), while a parallel turn draws the skis together and rolls
-    // them onto their edges into the turn (angulation) — visually distinct from the
-    // wedge. Purely cosmetic; none of this touches the physics.
-    if (snowman.userData && snowman.userData.leftSki && snowman.userData.rightSki) {
-      const ls = snowman.userData.leftSki, rs = snowman.userData.rightSki;
-      const lerp = Math.min(1, delta * 10);
-      const wedge = snowplow ? 0.35 : 0.0; // radians; tips angled inward
-      ls.rotation.y += ((-wedge) - ls.rotation.y) * lerp;
-      rs.rotation.y += ((wedge) - rs.rotation.y) * lerp;
-      // Parallel angulation: both skis edge the same way (into the turn) and slide
-      // toward each other; everything relaxes back to neutral otherwise.
-      const isParallel = technique === 'parallel';
-      const edge = isParallel ? steering * 0.28 : 0.0;
-      const draw = isParallel ? 0.35 : 0.0;
-      const lbx = (snowman.userData.leftSkiBaseX ?? -1) + draw;
-      const rbx = (snowman.userData.rightSkiBaseX ?? 1) - draw;
-      ls.rotation.z += (edge - ls.rotation.z) * lerp;
-      rs.rotation.z += (edge - rs.rotation.z) * lerp;
-      ls.position.x += (lbx - ls.position.x) * lerp;
-      rs.position.x += (rbx - rs.position.x) * lerp;
-      snowman.userData.technique = technique;
-    }
-    
-    // Update y position to terrain height when not in air
-    pos.y = terrainHeightAtPosition;
-  }
-  
-  // Apply velocity to position
-  pos.x += velocity.x * delta;
-  pos.z += velocity.z * delta;
-  
-  // Store current terrain height for next frame
-  lastTerrainHeight = terrainHeightAtPosition;
-  
-  // Update snowman position and rotation
-  snowman.position.set(pos.x, pos.y, pos.z);
-  
-  // More sophisticated rotation handling with smoothing
-  const movementDir = { x: velocity.x, z: velocity.z };
-  
-  // Add persistent rotation value for smoother transitions
-  if (!snowman.userData) snowman.userData = {};
-  if (snowman.userData.targetRotationY === undefined) {
-    snowman.userData.targetRotationY = Math.PI; // Default facing downhill
-    snowman.userData.currentRotX = 0;
-    snowman.userData.currentRotZ = 0;
-  }
-  
-  if (currentSpeed > 0.5) { // Only rotate if moving with significant speed
-    // Calculate target rotation - keep existing if below threshold
-    snowman.userData.targetRotationY = Math.atan2(movementDir.x, movementDir.z);
-  }
-  
-  // Apply rotation with smoothing - more stability at higher speeds
-  const rotationSmoothingY = Math.min(1, Math.max(0.05, delta * 2.5));
-  
-  // Interpolate toward target rotation - never rotate more than 15 degrees at once
-  const currentRotY = snowman.rotation.y;
-  const targetRotY = snowman.userData.targetRotationY;
-  
-  // Determine the shortest rotation direction (handle wrapping at 2π)
-  let deltaRotation = targetRotY - currentRotY;
-  if (deltaRotation > Math.PI) deltaRotation -= Math.PI * 2;
-  if (deltaRotation < -Math.PI) deltaRotation += Math.PI * 2;
-  
-  // Apply smoothed rotation with a max rate to prevent spinning
-  const maxRotationRate = delta * 3; // Max ~180° per second
-  const appliedDelta = Math.max(-maxRotationRate, Math.min(maxRotationRate, deltaRotation * rotationSmoothingY));
-  snowman.rotation.y += appliedDelta;
-  
-  // Normalize rotation to 0-2π range
-  snowman.rotation.y = snowman.rotation.y % (Math.PI * 2);
-  if (snowman.rotation.y < 0) snowman.rotation.y += Math.PI * 2;
-  
-  // Calculate a tilt based on the slope and turning with improved smoothing
-  // Use wider sample points for more stable gradients
-  const sampleDist = 0.4; // Even wider sampling for stability
-  const gradX = (getTerrainHeight(pos.x + sampleDist, pos.z) - getTerrainHeight(pos.x - sampleDist, pos.z)) / (2 * sampleDist);
-  const gradZ = (getTerrainHeight(pos.x, pos.z + sampleDist) - getTerrainHeight(pos.x, pos.z - sampleDist)) / (2 * sampleDist);
-  
-  // Add more dramatic jump rotation - lean forward during jumps
-  let jumpTilt = 0;
-  if (isInAir) {
-    // More dramatic tilt during jumps, especially on takeoff
-    if (verticalVelocity > 0) {
-      // Lean back on ascent
-      jumpTilt = -Math.min(0.4, verticalVelocity * 0.03);
-    } else {
-      // Lean forward on descent, more as you fall faster
-      jumpTilt = Math.min(0.5, -verticalVelocity * 0.025);
-    }
-  }
-  
-  // Add more controlled turning tilt with speed-based scaling
-  const turnTiltFactor = Math.min(0.3, currentSpeed / 25); // Less tilt at lower speeds
-  const turnTilt = velocity.x * turnTiltFactor;
-  
-  // Limit maximum tilt angles to prevent unrealistic leaning
-  const maxTiltAngle = 0.25; // Reduced to about 14 degrees maximum tilt
-  
-  // Apply smoothing and clamping to rotation values
-  const targetRotX = gradZ * 0.3 + jumpTilt; // Add jump tilt to X rotation
-  const targetRotZ = -gradX * 0.3 - turnTilt;
-  
-  // Significantly improve tilt smoothing
-  const tiltSmoothing = isInAir ? 0.05 : 0.08; // Lower values for smoother transitions
-  
-  // Use lerp for smooth rotation
-  snowman.userData.currentRotX = snowman.userData.currentRotX || 0;
-  snowman.userData.currentRotZ = snowman.userData.currentRotZ || 0;
-  
-  // Smoothly transition current rotations toward target
-  snowman.userData.currentRotX += (targetRotX - snowman.userData.currentRotX) * tiltSmoothing;
-  snowman.userData.currentRotZ += (targetRotZ - snowman.userData.currentRotZ) * tiltSmoothing;
-  
-  // Apply clamped rotations 
-  snowman.rotation.x = Math.max(-maxTiltAngle, Math.min(maxTiltAngle, snowman.userData.currentRotX));
-  snowman.rotation.z = Math.max(-maxTiltAngle, Math.min(maxTiltAngle, snowman.userData.currentRotZ));
-  
-  // Check if snowman is off the terrain or falling
-  const fallThreshold = 0.5; // How far below terrain to allow before reset
-  
-  // Check for tree collisions
-  // Use the window variable for collision radius if set (for testing), otherwise use default
-  const treeCollisionRadius = window.treeCollisionRadius || 2.5; // Collision distance for trees
-  
-  // In test mode, output complete tree positions for debugging
-  if (window.location.search.includes('test=true') && treePositions.length > 0) {
-    console.log(`TREES LOADED: ${treePositions.length} trees found`);
-    console.log(`SNOWMAN POS: x=${pos.x.toFixed(2)}, y=${pos.y.toFixed(2)}, z=${pos.z.toFixed(2)}`);
-    console.log(`FIRST TREE: x=${treePositions[0].x.toFixed(2)}, y=${treePositions[0].y.toFixed(2)}, z=${treePositions[0].z.toFixed(2)}`);
-  }
-  
-  // Count trees in extended terrain area for logging
-  const extendedTreesCount = treePositions.filter(t => t.z < -80).length;
-  const totalTreesCount = treePositions.length;
-  
-  // Log tree information when in test mode
-  if (window.location.search.includes('test=true') && !window._treeCheckLogged) {
-    console.log(`TREE COLLISION INFO: ${totalTreesCount} total trees, ${extendedTreesCount} in extended area (z < -80)`);
-    
-    // Log the ranges to verify coverage
-    if (treePositions.length > 0) {
-      const zMin = Math.min(...treePositions.map(t => t.z));
-      const zMax = Math.max(...treePositions.map(t => t.z));
-      const xMin = Math.min(...treePositions.map(t => t.x));
-      const xMax = Math.max(...treePositions.map(t => t.x));
-      console.log(`TREE COLLISION RANGES: X: ${xMin.toFixed(1)} to ${xMax.toFixed(1)}, Z: ${zMin.toFixed(1)} to ${zMax.toFixed(1)}`);
-    }
-    window._treeCheckLogged = true;
-  }
-  
-  // Check collision with any tree
-  const treeCollision = treePositions.some(treePos => {
-    // Special case for tests - direct position match or very close positions always collide
-    // Use a small epsilon for floating point comparison, increased for test reliability
-    // This helps with floating-point precision issues in tests
-    const epsilon = window.location.search.includes('test') ? 0.1 : 0.001;
-    
-    const exactMatch = 
-      Math.abs(pos.x - treePos.x) < epsilon && 
-      Math.abs(pos.z - treePos.z) < epsilon;
-    
-    if (exactMatch) {
-      if (window.location.search.includes('test=true')) {
-        console.log(`DIRECT TREE HIT at (${pos.x.toFixed(2)}, ${pos.z.toFixed(2)})`);
-      }
-      return true;
-    }
-    
-    // Check horizontal distance for collision (2D distance ignoring height)
-    const dx = pos.x - treePos.x;
-    const dz = pos.z - treePos.z;
-    const horizontalDistance = Math.sqrt(dx*dx + dz*dz);
-    
-    // We only detect collision if the horizontal distance is close enough
-    // Tree collision only happens when snowman is on the ground or close to it
-    const isCloseEnough = horizontalDistance < treeCollisionRadius;
-    
-    // Only consider jumping over trees when genuinely in the air AND moving upward AND high enough
-    const isJumpingHighAboveTrees = isInAir && verticalVelocity > 0 && pos.y > (treePos.y + 5);
-    
-    // Debug collision in browser tests when needed
-    if (window.location.search.includes('test=true')) {
-      // For extended terrain trees, log additional info
-      const inExtendedArea = treePos.z < -80;
-      console.log(`TREE CHECK: dist=${horizontalDistance.toFixed(2)}, radius=${treeCollisionRadius}, jumping=${isJumpingHighAboveTrees}, collision=${isCloseEnough && !isJumpingHighAboveTrees}, extended=${inExtendedArea}`);
-      
-      // Extra debugging info for very close trees
-      if (horizontalDistance < 5) {
-        console.log(`CLOSE TREE: x=${treePos.x.toFixed(2)}, y=${treePos.y.toFixed(2)}, z=${treePos.z.toFixed(2)}, snowman: x=${pos.x.toFixed(2)}, y=${pos.y.toFixed(2)}, z=${pos.z.toFixed(2)}`);
-      }
-    }
-    
-    // Special handling for tests in browser-tests.js
-    if (window.location.search.includes('test') && horizontalDistance < 0.5) {
-      console.log(`TEST MODE: Forcing collision with very close tree (${horizontalDistance.toFixed(2)})`);
-      return true;
-    }
-    
-    // Special handling for tree jumping test
-    if (window.testTreeJumpingCheck && isJumpingHighAboveTrees) {
-      console.log(`TREE JUMPING TEST: Allowing jump over tree (dist=${horizontalDistance.toFixed(2)})`);
-      // Don't detect collision during jump - normal game behavior
-      return false;
-    }
-    
-    // Allow jumping over trees but collide when on the ground
-    return isCloseEnough && !isJumpingHighAboveTrees;
-  });
-
-  // Check collision with large, exposed rocks. Small half-buried stones remain
-  // terrain detail; only positions returned by Mountains.addRocks reach this list.
-  const rockCollision = rockPositions.some(rockPos => {
-    const dx = pos.x - rockPos.x;
-    const dz = pos.z - rockPos.z;
-    const horizontalDistance = Math.sqrt(dx*dx + dz*dz);
-    // Collision radius (max 3u). Kept in sync with Mountains.rockCollisionRadius,
-    // which the placement-time safe-zone uses to exclude rocks that would reach the
-    // ski lane/spawn pocket. Duplicated inline (rather than imported) because
-    // snowman.ts must stay free of relative imports: some Node test harnesses load
-    // it without the .js->.ts resolve hook, and Node 22 won't map ./mountains.js.
-    const rockRadius = Math.max(1.25, Math.min(3.0, rockPos.size * 0.75 + 0.75));
-    const exposedRockTop = rockPos.y + rockPos.size * 0.7;
-    // Clearance is height-based: once the snowman is airborne and above the rock
-    // top it clears the hazard whether it is still rising or already descending past
-    // the jump apex. (Requiring upward motion made descending-but-high jumps crash.)
-    const isJumpingOverRock = isInAir && pos.y > exposedRockTop + 0.5;
-
-    if (window.location.search.includes('test=true') && horizontalDistance < 5) {
-      console.log(`ROCK CHECK: dist=${horizontalDistance.toFixed(2)}, radius=${rockRadius.toFixed(2)}, jumping=${isJumpingOverRock}, collision=${horizontalDistance < rockRadius && !isJumpingOverRock}`);
-    }
-
-    return horizontalDistance < rockRadius && !isJumpingOverRock;
-  });
-  
-  // Reset if: reaches end of slope, goes off sides, falls off terrain, or hits a tree
-  // Allow wider boundaries to match the extended mountain terrain
-  // Only skip boundary check during regression/tree tests, but NOT during browser tests or unified tests
-  const inExtendedMountainTest = window.location.search.includes('test=regression') || 
-                                window.location.search.includes('test=tree'); // Only skip for specific tests
-  if (pos.z < -195 || // Extended from -95 to -195 for longer run
-      (!inExtendedMountainTest && Math.abs(pos.x) > 120) || // Keep boundary check during browser/unified tests
-      (!isInAir && pos.y < terrainHeightAtPosition - fallThreshold) ||
-      treeCollision ||
-      rockCollision) {
-    
-    if (gameActive) {
-      // Determine the reason for game over
-      let reason = "You crashed!";
-      
-      if (treeCollision) {
-        reason = "BANG!!! You hit a tree!";
-      } else if (rockCollision) {
-        reason = "BANG!!! You hit a rock!";
-      } else if (pos.z < -195) {
-        reason = "You reached the end of the slope!";
-      } else if (Math.abs(pos.x) > 120) {
-        reason = "You went off the mountain!";
-      } else if (!isInAir && pos.y < terrainHeightAtPosition - fallThreshold) {
-        reason = "You fell off the terrain!";
-      }
-      
-      showGameOver(reason);
-    }
-  }
-  
-  // Return updated state variables
-  return {
-    isInAir,
-    verticalVelocity,
-    lastTerrainHeight,
-    airTime,
-    jumpCooldown,
-    turnPhase,
-    currentTurnDirection,
-    turnChangeCooldown,
-    currentSpeed,
-    technique,
-    justLanded,
-    landingForce
-  };
-}
-
-// Add test hook functions for tree collision testing
-function addTestHooks(pos: PlayerPos, showGameOver: ShowGameOverFn, getTerrainHeight: TerrainHeightFn) {
-  console.log("Snowman.addTestHooks called - setting up test hooks");
-  
-  if (!window.testHooks) {
-    window.testHooks = {};
-  }
-  
-  // Add a force collision function that can be called by tests
-  window.testHooks.forceTreeCollision = function() {
-    console.log("TEST: forceTreeCollision hook called");
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Call the function directly to ensure it works
-    console.log("TEST: Forcing tree collision (direct call)");
-    try {
-      directShowGameOver("BANG!!! You hit a tree!");
-      console.log("TEST: Successfully called showGameOver");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver:", error);
-      // As a fallback, just simulate the collision for the test
-      window.testCollisionDetected = true;
-    }
-    
-    return true;
-  };
-  
-  // Add a tree collision checking function
-  window.testHooks.checkTreeCollision = function(x: number, z: number) {
-    console.log("TEST: checkTreeCollision hook called");
-    // Create a test tree at the specified position
-    const testTree = { 
-      x: x, 
-      y: getTerrainHeight(x, z), 
-      z: z 
-    };
-    
-    // For tree jumping test, check if we're in the air
-    const isJumpingHighAboveTrees = isInAir && verticalVelocity > 0 && pos.y > (testTree.y + 5);
-    console.log(`TEST: Jump check - isInAir=${isInAir}, verticalVelocity=${verticalVelocity}, pos.y=${pos.y}, tree.y=${testTree.y}, jumping=${isJumpingHighAboveTrees}`);
-    
-    // If we're testing tree jumping, we need to handle it properly
-    if (window.testTreeJumpingCheck) {
-      // This is the tree jumping test
-      window.testTreeJumpingCheck = false; // Reset the flag
-      
-      // We should detect the collision even when jumping in the test hook
-      console.log("TEST: checkTreeCollision for jumping test - will detect collision regardless of jumping state");
-      
-      // Create a direct function reference to ensure it works
-      const directShowGameOver = window.showGameOver || showGameOver;
-      
-      // Always detect collision in the jumping test to verify the hook works
-      try {
-        directShowGameOver("BANG!!! You hit a tree (ignoring jump)!");
-        console.log("TEST: Successfully called showGameOver from checkTreeCollision (jumping test)");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver from checkTreeCollision (jumping test):", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // For regular testing, respect the jumping logic
-    if (isJumpingHighAboveTrees) {
-      console.log("TEST: Snowman is jumping high above trees - no collision");
-      return false;
-    }
-    
-    // Position the snowman directly at the tree for collision testing
-    pos.x = x;
-    pos.z = z;
-    pos.y = testTree.y;
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Check for collision with this test tree
-    console.log(`TEST: Checking collision at (${x}, ${z})`);
-    
-    // Always detect collision in test regardless of distance
-    console.log("TEST: Tree collision detected");
-    try {
-      directShowGameOver("BANG!!! You hit a tree!");
-      console.log("TEST: Successfully called showGameOver from checkTreeCollision");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver from checkTreeCollision:", error);
-      // As a fallback, just simulate the collision for the test
-      window.testCollisionDetected = true;
-    }
-    return true;
-  };
-  
-  // Add a function to test collision in extended terrain area
-  window.testHooks.checkExtendedTerrainCollision = function() {
-    console.log("TEST: checkExtendedTerrainCollision hook called");
-    
-    // Create a direct function reference to ensure it works
-    const directShowGameOver = window.showGameOver || showGameOver;
-    
-    // Check if we have any trees in extended terrain
-    if (!window.treePositions || !window.treePositions.length) {
-      console.log("TEST: No trees available for extended terrain test");
-      // Still show collision for test to pass
-      try {
-        directShowGameOver("BANG!!! You hit a tree in extended terrain (simulated)!");
-        console.log("TEST: Successfully called showGameOver for extended terrain");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver for extended terrain:", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // Find a tree in extended terrain (z < -80)
-    const extendedTrees = window.treePositions.filter((t: { z: number }) => t.z < -80);
-    if (extendedTrees.length === 0) {
-      console.log("TEST: No trees found in extended terrain (z < -80)");
-      // Still show collision for test to pass
-      try {
-        directShowGameOver("BANG!!! You hit a tree in extended terrain (simulated)!");
-        console.log("TEST: Successfully called showGameOver for extended terrain");
-      } catch (error) {
-        console.error("TEST ERROR: Failed to call showGameOver for extended terrain:", error);
-        window.testCollisionDetected = true;
-      }
-      return true;
-    }
-    
-    // Use the first tree in extended terrain for testing
-    const testTree = extendedTrees[0];
-    console.log(`TEST: Using tree at (${testTree.x.toFixed(1)}, ${testTree.z.toFixed(1)}) in extended terrain`);
-    
-    // Position snowman at the tree for collision
-    pos.x = testTree.x;
-    pos.z = testTree.z;
-    pos.y = testTree.y;
-    
-    // Check for collision 
-    console.log("TEST: Positioned snowman directly on extended terrain tree, checking collision");
-    
-    // Always trigger collision in test
-    try {
-      directShowGameOver("BANG!!! You hit a tree in extended terrain!");
-      console.log("TEST: Successfully called showGameOver for extended terrain collision");
-    } catch (error) {
-      console.error("TEST ERROR: Failed to call showGameOver for extended terrain collision:", error);
-      window.testCollisionDetected = true;
-    }
-    return true;
-  };
-  
-  // For debugging in tests
-  console.log("Snowman test hooks installed: forceTreeCollision, checkTreeCollision, checkExtendedTerrainCollision");
-}
-
-// Export snowman functions
-export const Snowman = {
-  createSnowman,
-  resetSnowman,
-  updateSnowman,
-  addTestHooks
-};
-
-// Snowman is imported directly by snowglider.js and the gameplay browser tests
-// (issue #84).
+// `export *` re-exports the `Snowman` object AND every contract type (PlayerPos,
+// UpdateResult, …) that physics.ts / hud.ts consume.
+//
+// NOTE: the physics-invariant harness (tests/verification/physics_invariant_harness.js)
+// runs under bare Node with NO `.js`->`.ts` resolve hook, so it cannot resolve this
+// facade's `./snowman/index.js` re-export. It imports the relocated implementation
+// directly (`src/snowman/index.ts`) instead; keep that harness import in sync if this
+// facade ever changes. `src/snowman/index.ts` stays free of *relative* imports so it
+// remains loadable under that bare-Node harness.
+export * from './snowman/index.js';
